Move list mode into conversation list controller
