+++ app/src/main/scala/com/waz/zclient/conversationlist/ConversationListFragment.scala
@@ -34,7 +34,8 @@ import com.waz.utils.returning
 import com.waz.zclient.common.controllers.UserAccountsController
 import com.waz.zclient.common.controllers.global.AccentColorController
 import com.waz.zclient.conversation.ConversationController
-import com.waz.zclient.conversationlist.adapters.{ArchiveConversationListAdapter, ConversationListAdapter, NormalConversationListAdapter}
+import com.waz.zclient.conversationlist.ConversationListController.{Archive, ListMode, Normal}
+import com.waz.zclient.conversationlist.adapters.{ArchiveConversationListAdapter, NormalConversationListAdapter}
 import com.waz.zclient.conversationlist.views.{ArchiveTopToolbar, ConversationListTopToolbar, NormalTopToolbar}
 import com.waz.zclient.core.stores.conversation.ConversationChangeRequester
 import com.waz.zclient.log.LogUI._
@@ -70,12 +71,12 @@ abstract class ConversationListFragment extends BaseFragment[ConversationListFra
   lazy val convListController     = inject[ConversationListController]
 
   protected var subs = Set.empty[Subscription]
-  protected val adapterMode: ConversationListAdapter.ListMode
+  protected val adapterMode: ListMode
 
   protected lazy val topToolbar: ViewHolder[_ <: ConversationListTopToolbar] = view[ConversationListTopToolbar](R.id.conversation_list_top_toolbar)
 
   lazy val adapter = returning(adapterMode match {
-    case ConversationListAdapter.Normal =>
+    case Normal =>
       returning(new NormalConversationListAdapter) { a =>
         val dataSource = for {
           regular  <- convListController.regularConversationListData
@@ -86,7 +87,7 @@ abstract class ConversationListFragment extends BaseFragment[ConversationListFra
           a.setData(regular, incoming)
         }
       }
-    case ConversationListAdapter.Archive =>
+    case Archive =>
       returning(new ArchiveConversationListAdapter) { a =>
         convListController.archiveConversationListData.onUi { archive =>
           a.setData(archive)
@@ -158,7 +159,7 @@ class ArchiveListFragment extends ConversationListFragment with OnBackPressedLis
 
   override val layoutId = R.layout.fragment_archive_list
   override lazy val topToolbar = view[ArchiveTopToolbar](R.id.conversation_list_top_toolbar)
-  override protected val adapterMode = ConversationListAdapter.Archive
+  override protected val adapterMode = Archive
 
   override def onViewCreated(view: View, savedInstanceState: Bundle) = {
     super.onViewCreated(view, savedInstanceState)
@@ -178,7 +179,7 @@ object NormalConversationListFragment {
 class NormalConversationFragment extends ConversationListFragment {
 
   override val layoutId = R.layout.fragment_conversation_list
-  override protected val adapterMode = ConversationListAdapter.Normal
+  override protected val adapterMode = Normal
 
   lazy val zms = inject[Signal[ZMessaging]]
   lazy val accentColor = inject[AccentColorController].accentColor
